refactor filtering utils
