+++ js/ui/grid_core/ui.grid_core.utils.js
@@ -1,7 +1,7 @@
 import { getHeight } from '../../core/utils/size';
 import $ from '../../core/renderer';
-import { isDefined, isFunction, isString } from '../../core/utils/type';
-import { when, Deferred } from '../../core/utils/deferred';
+import { isDefined, isFunction } from '../../core/utils/type';
+import { when } from '../../core/utils/deferred';
 import sharedFiltering from '../shared/filtering';
 import { format } from '../../core/utils/string';
 import { each } from '../../core/utils/iterator';
@@ -14,9 +14,6 @@ import { normalizeSortingInfo as normalizeSortingInfoUtility } from '../../data/
 import formatHelper from '../../format_helper';
 import { getWindow } from '../../core/utils/window';
 import eventsEngine from '../../events/core/events_engine';
-import { DataSource } from '../../data/data_source/data_source';
-import { normalizeDataSourceOptions } from '../../data/data_source/utils';
-import variableWrapper from '../../core/utils/variable_wrapper';
 
 const DATAGRID_SELECTION_DISABLED_CLASS = 'dx-selection-disabled';
 const DATAGRID_GROUP_OPENED_CLASS = 'dx-datagrid-group-opened';
@@ -152,23 +149,6 @@ const equalFilterParameters = function(filter1, filter2) {
         return toComparable(filter1) == toComparable(filter2); // eslint-disable-line eqeqeq
     }
 };
-
-function normalizeGroupingLoadOptions(group) {
-    if(!Array.isArray(group)) {
-        group = [group];
-    }
-
-    return group.map((item, i) => {
-        if(isString(item)) {
-            return {
-                selector: item,
-                isExpanded: i < group.length - 1,
-            };
-        }
-
-        return item;
-    });
-}
 
 export default {
     renderNoDataText: function($element) {
@@ -566,112 +546,5 @@ export default {
         }
 
         return 15000000 / this.getPixelRatio(getWindow());
-    },
-
-    normalizeLookupDataSource(lookup) {
-        let lookupDataSourceOptions;
-        if(lookup.items) {
-            lookupDataSourceOptions = lookup.items;
-        } else {
-            lookupDataSourceOptions = lookup.dataSource;
-            if(isFunction(lookupDataSourceOptions) && !variableWrapper.isWrapped(lookupDataSourceOptions)) {
-                lookupDataSourceOptions = lookupDataSourceOptions({});
-            }
-        }
-
-        return normalizeDataSourceOptions(lookupDataSourceOptions);
-    },
-
-    getWrappedLookupDataSource(column, dataSource, filter) {
-        if(!dataSource) {
-            return [];
-        }
-
-        const lookupDataSourceOptions = this.normalizeLookupDataSource(column.lookup);
-
-        if(column.calculateCellValue !== column.defaultCalculateCellValue) {
-            return lookupDataSourceOptions;
-        }
-
-        const hasGroupPaging = dataSource.remoteOperations().groupPaging;
-        const hasLookupOptimization = column.displayField && isString(column.displayField);
-
-        let cachedUniqueRelevantItems;
-
-        const sliceItems = (items, loadOptions) => {
-            const start = loadOptions.skip ?? 0;
-            const end = loadOptions.take ? start + loadOptions.take : items.length;
-            return items.slice(start, end);
-        };
-
-        const loadUniqueRelevantItems = (loadOptions) => {
-            const group = normalizeGroupingLoadOptions(
-                hasLookupOptimization ? [column.dataField, column.displayField] : column.dataField
-            );
-            const d = new Deferred();
-
-            if(!hasGroupPaging && cachedUniqueRelevantItems) {
-                d.resolve(sliceItems(cachedUniqueRelevantItems, loadOptions));
-            } else {
-                dataSource.load({
-                    filter,
-                    group,
-                    take: hasGroupPaging ? loadOptions.take : undefined,
-                    skip: hasGroupPaging ? loadOptions.skip : undefined,
-                }).done((items) => {
-                    cachedUniqueRelevantItems = items;
-                    d.resolve(hasGroupPaging ? items : sliceItems(items, loadOptions));
-                }).fail(d.fail);
-            }
-
-            return d;
-        };
-
-        const lookupDataSource = {
-            ...lookupDataSourceOptions,
-            __dataGridSourceFilter: filter,
-            load: (loadOptions) => {
-                const d = new Deferred();
-                loadUniqueRelevantItems(loadOptions).done((items) => {
-                    if(items.length === 0) {
-                        d.resolve([]);
-                        return;
-                    }
-
-                    const filter = this.combineFilters(
-                        items.flatMap((data) => data.key).map((key => [
-                            column.lookup.valueExpr, key,
-                        ])),
-                        'or'
-                    );
-
-                    const newDataSource = new DataSource({
-                        ...lookupDataSourceOptions,
-                        ...loadOptions,
-                        filter: this.combineFilters([filter, loadOptions.filter], 'and'),
-                        paginate: false, // pagination is included to filter
-                    });
-
-                    newDataSource
-                        .load()
-                        .done(d.resolve)
-                        .fail(d.fail);
-                }).fail(d.fail);
-                return d;
-            },
-            key: column.lookup.valueExpr,
-            byKey(key) {
-                const d = Deferred();
-                this.load({
-                    filter: [column.lookup.valueExpr, '=', key],
-                }).done(arr => {
-                    d.resolve(arr[0]);
-                });
-
-                return d.promise();
-            },
-        };
-
-        return lookupDataSource;
     },
 };
